Add new features for expense management and refueling tracking
- Implemented IconWithText widget for better UI representation.
- Created RefuelingModel to manage refueling data with JSON serialization.
- Developed VehicleModel to handle vehicle information with JSON serialization.
- Added CreateExpenseBinding for dependency injection of CreateExpenseController.
- Built CreateExpenseController to manage the expense creation logic, including form handling and data validation.
- Designed CreateExpenseView for user interface to input expense details, including date, time, odometer, and attachments.
- Integrated image picker functionality for attaching files to expenses.
- Enhanced user experience with localized text and improved UI elements.

+++ lib/model/vehicle/vehicle_model.dart
@@ -37,7 +37,7 @@ class VehicleModel {
     manufacturer = json["manufacturer"] ?? "";
     model = json["model"] ?? "";
     licensePlate = json["license_plate"] ?? "";
-    year = json["year"] ?? 0;
+    year = (json["year"] ?? 0).toInt();
     tankConfiguration = json["tank_configuration"] ?? "";
     fuelType = json["fuel_type"] ?? "";
     fuelCapacity = json["fuel_capacity"] ?? "";
